Do not clear cache on new path before turning maintenance mode on
We check to see if the locale module is enabled to see if we should
import translations. Before we did this by executing the command to see
if it's enabled and based on the output we added the translation update
command. Part of the command of checking for the module, was clearing
the cache. So now, instead of running the command and building the
commands that we will be running later on based on the output, we just
make the command (that checks for the module) part of the command chain,
and chain them together with '&&'.

The issue with the old way we did it was that we cleared the cache in
the new codebase and this was executed _before_ we enabled maintenance
mode (because the command for checking for the locale module
was executed immediately, and not part of the chain we were building).
So when we enabled maintenance mode in the old codebase, we did it while
loading cache from the new codebase. This resulted in Drupal trying to
look form plugins, classes, ... from the new codebase in the old one,
which, as you can imagine, resulted in errors in some edge cases.