Add options to the mysqldump command
--no-tablespaces
From the mysqldump manual at
https://dev.mysql.com/doc/refman/8.0/en/mysqldump.html:

mysqldump requires at least the SELECT privilege for dumped tables, SHOW
VIEW for dumped views, TRIGGER for dumped triggers, LOCK TABLES if the
--single-transaction option is not used, and (as of MySQL 8.0.21)
PROCESS if the --no-tablespaces option is not used. Certain options
might require other privileges as noted in the option descriptions.

--skip-add-locks
We're in the process of moving to a new environment. This environment's
database layer is a percona xtradb cluster, which doesn't allow locking
tables. This means we have to pass the `--skip-add-locks` option to the
`mysqldump` command. We know we'll take a performance hit by adding this,
but it seems it's our only option for now.